% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyPath.R
\name{adjacencyPath}
\alias{adjacencyPath}
\title{Find adjectives and the nouns they describe}
\usage{
adjacencyPath(Sentence, Path = "amod")
}
\arguments{
\item{Sentence}{a record (row) in the GeoDeepDive NLP output}

\item{Path}{a stanford core nlp pose code}
}
\value{
A character matrix
}
\description{
Returns a tuple of adjectives and nouns in a target sentence.
}
\details{

}
\examples{

# TBDDDDD

}
\author{
Andrew A. Zaffos & Erika T. Ito
}
