% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPunctuation.R
\name{cleanPunctuation}
\alias{cleanPunctuation}
\title{Clean or replace punctuation oddities}
\usage{
cleanPunctuation(Sentence)
}
\arguments{
\item{Sentence}{A string or vector of strings}
}
\value{
A string or vector of strings
}
\description{
A function to help R better parse comma separated arrays taken from PostgreSQL. This includes cleaning of brackets, spaces, commas, and parenthesis.
}
\details{
The punctuation changes made here are not related to GeoDeepDive or its NLP output, rather it is solving artefacts of how R handles concatenated arrays in PostgreSQL. For example, PostgreSQL brackets its arrayws with { }, which R keeps when reading the array in, which is undesirable.
}
\examples{

# Example

}
\author{
Andrew A. Zaffos & Erika T. Ito
}
