% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutivePoses.R
\name{consecutivePoses}
\alias{consecutivePoses}
\title{Find clusters of consecutive posees}
\usage{
consecutivePoses(Sentence, Pose = "NNP")
}
\arguments{
\item{Sentence}{a GeoDeepDive output nlp output record}

\item{Pose}{a string}
}
\value{
A character matrix
}
\description{
Returns tuples of GeoDeepDive docid, sentid, and sets of consecutive parts of speech (i.e., poses).
}
\details{
This function will find groupings of parts of speech that occur consecutively within a sentence using the poses output of the StanfordCoreNLP. It primarily makes sense to look for clusters of proper nouns ("NNP") to extract entities with multi-word names (e.g., people, places, organizations). However, you can look for other types of pose as well. The format is a matrix of tuples of GeoDeepDive docid, sentid, and the cluster of poses.
}
\examples{

# TBDDDDD

}
\author{
Andrew A. Zaffos & Erika T. Ito
}
