% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGDD.R
\name{downloadGDD}
\alias{downloadGDD}
\title{Fetch Article Metadata}
\usage{
downloadGDD(Journal = "", Term = "", DocID = "", Surname = "",
  Given = "")
}
\arguments{
\item{Journal}{a character vector}

\item{Term}{a character vector}

\item{DocID}{a character vector}

\item{Surname}{a character vector; author's last name}

\item{Given}{a character vector; author's first name}
}
\value{
A list of article metadata
}
\description{
Downloads article bibjson metadata from \code{/articles} route of the GeoDeepDive API.
}
\details{
This function allows you to download article metadata from the GeoDeepDive library. It allows you
to choose article metadata based on one more parameters including the journal title, GeoDeepDive assigned
document id, author first name, author last name, or term within the article. This data is pulled from the
\href{https://geodeepdive.org/api/articles}{GeoDeepDive /articles} API route

Note: There are other API parameter options that are not included in this simplified wrapper.
}
\examples{
# Download the metadata for all articles from the Journal of Vertebrate Paleontology
GDDMetadata<-downloadGDD(Journal="Journal of Vertebrate Paleontology")

# Download the metadata for all articles from the Journal of Vertebrate Paleontology written by Michael J. Benton
GDDMetadata<-downloadGDD(Journal="Journal of Vertebrate Paleontology", Surname="Benton", Given="Michael")

}
\author{
Andrew A. Zaffos & Erika T. Ito
}
