% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSentence.R
\name{parseSentence}
\alias{parseSentence}
\title{A function to make GDD NLP output human-readable}
\usage{
parseSentence(Sentence, Parameters = c("words", "dep_paths",
  "dep_parents"))
}
\arguments{
\item{Sentence}{a record (row) in the GDD NLP output}

\item{Parameters}{a vector of sentence properties}
}
\value{
A matrix
}
\description{
This function will convert a sentence from the raw GeoDeepDive nlp output tables into a more human-readable format.
}
\details{
This function will convert a sentence from the raw GeoDeepDive nlp output tables into a more human-readable format. It returns a matrix of where each column represents an individual word in the sentence, and each row represents a specific property of the word - e.g., poses, dep_parents, dep_paths.
}
\examples{

# TBDDDDD

}
\author{
Andrew A. Zaffos & Erika T. Ito
}
