% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNGRAM.R
\name{plotNGRAM}
\alias{plotNGRAM}
\title{Plot Article Age Data}
\usage{
plotNGRAM(Term, Publisher, Journal)
}
\arguments{
\item{Term}{a character string}

\item{Publisher}{a character vector}

\item{Journal}{a character vector}
}
\value{
A time series plot
}
\description{
Plots the age distribution of articles using certain terms of interest (i.e. ngram).
}
\details{
This function plots the age distribution of articles from the GeoDeepDive digital library that contain a certain term of interest. This
data comes from the \href{https://geodeepdive.org/api/articles}{GeoDeepDive /articles} API route.
}
\examples{

# TBDDDDD

}
\author{
Andrew A. Zaffos & Erika T. Ito
}

